4.0-hello-fastapi/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Antti", email = "antti.linno@gmail.com"},
]
dependencies = [
    "uvicorn>=0.17.6",
    "fastapi>=0.78.0",
    "pydantic>=1.9.0",
    "httpx>=0.22.0",
    "pyyaml>=6.0",
    "requests>=2.27.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
